backend/requirements.txt
fastapi[standard]==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pydantic-settings==2.*
google-genai==1.*
celery[redis]==5.*
redis==5.*
boto3==1.*
python-jose[cryptography]==3.*
httpx==0.28.*
fal-client==0.*
Pillow==11.*
scikit-learn==1.*
numpy==2.*
pytest==8.*
pytest-asyncio==0.*
sentry-sdk[fastapi]==2.*
